proper response codes for rest calls

+++ scase-wp5-asset-registry/scase-wp5-asset-registry-backend-impl/src/main/java/eu/scasefp7/assetregistry/rest/ProjectResource.java
@@ -1,8 +1,17 @@
 package eu.scasefp7.assetregistry.rest;
 
-import static eu.scasefp7.assetregistry.rest.ResourceTools.redirect;
-
-import java.util.List;
+import eu.scasefp7.assetregistry.data.Project;
+import eu.scasefp7.assetregistry.dto.JsonProject;
+import eu.scasefp7.assetregistry.dto.ProjectDTO;
+import eu.scasefp7.assetregistry.service.ProjectService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import org.apache.http.HttpStatus;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
@@ -17,34 +26,21 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-
-import org.apache.http.HttpStatus;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import eu.scasefp7.assetregistry.data.Project;
-import eu.scasefp7.assetregistry.dto.JsonProject;
-import eu.scasefp7.assetregistry.dto.ProjectDTO;
-import eu.scasefp7.assetregistry.service.ProjectService;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
+import java.util.List;
 
 /**
  * rest api for a project.
- * @author rmagnus
  *
+ * @author rmagnus
  */
-@Path(AssetRegistryRestApp.PART_PROJECT)
-@Api(value = AssetRegistryRestApp.PART_PROJECT, description = "provides projects")
-@Produces("application/json;charset=UTF-8")
-@Consumes("application/json")
+@Path( AssetRegistryRestApp.PART_PROJECT )
+@Api( value = AssetRegistryRestApp.PART_PROJECT, description = "provides projects" )
+@Produces( "application/json;charset=UTF-8" )
+@Consumes( "application/json" )
 @Stateless
-public class ProjectResource
-{
-    private static final Logger LOG = LoggerFactory.getLogger(ProjectResource.class);
+public class ProjectResource {
+
+    private static final Logger LOG = LoggerFactory.getLogger( ProjectResource.class );
 
     @EJB
     private ProjectService projectService;
@@ -56,22 +52,21 @@ public class ProjectResource
      * @return Project project
      */
     @GET
-    @Path("{id}")
-    @ApiOperation(value = "Finds a project in the repository by ID")
-    @ApiResponses(value = {
-            @ApiResponse(code = HttpStatus.SC_NO_CONTENT, message = "No content"),
-            @ApiResponse(code = HttpStatus.SC_BAD_REQUEST, message = "Request incorrect"),
-            @ApiResponse(code = HttpStatus.SC_NOT_FOUND, message = "Not found"),
-            @ApiResponse(code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error")})
-    public JsonProject get(@PathParam("id") @ApiParam(value = "project ID") long id)
-    {
-        JsonProject jsonProject = null;
+    @Path( "{id}" )
+    @ApiOperation( value = "Finds a project in the repository by ID",
+            response = JsonProject.class )
+    @ApiResponses( value = {
+            @ApiResponse( code = HttpStatus.SC_OK, message = "OK" ),
+            @ApiResponse( code = HttpStatus.SC_NO_CONTENT, message = "No content" ),
+            @ApiResponse( code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error" )} )
+    public Response get( @PathParam( "id" ) @ApiParam( value = "project ID" ) long id ) {
 
-        final Project projectEntity = this.projectService.find(id);
-        if(null!=projectEntity) {
-            jsonProject = this.projectService.convertEntityToJson(projectEntity);
+        final Project projectEntity = this.projectService.find( id );
+        if ( null != projectEntity ) {
+            JsonProject jsonProject = this.projectService.convertEntityToJson( projectEntity );
+            return Response.status( Response.Status.OK ).entity( jsonProject ).build();
         }
-        return jsonProject;
+        return Response.status( Response.Status.NO_CONTENT ).build();
     }
 
     /**
@@ -81,32 +76,23 @@ public class ProjectResource
      * @return Project project
      */
     @GET
-    @Path("{name}")
-    @ApiOperation(value = "Finds a project in the repository by its name")
-    @ApiResponses(value = {
-            @ApiResponse(code = HttpStatus.SC_NO_CONTENT, message = "No content"),
-            @ApiResponse(code = HttpStatus.SC_BAD_REQUEST, message = "Request incorrect"),
-            @ApiResponse(code = HttpStatus.SC_NOT_FOUND, message = "Not found"),
-            @ApiResponse(code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error")})
-    public JsonProject get(@PathParam("name") @ApiParam(value = "the name of the project") String name)
-    {
+    @Path( "{name}" )
+    @ApiOperation( value = "Finds a project in the repository by its name",
+            response = JsonProject.class )
+    @ApiResponses( value = {
+            @ApiResponse( code = HttpStatus.SC_OK, message = "OK" ),
+            @ApiResponse( code = HttpStatus.SC_NO_CONTENT, message = "No content" ),
+            @ApiResponse( code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error" )} )
+    public Response get( @PathParam( "name" ) @ApiParam( value = "the name of the project" ) String name ) {
 
-        Project projectEntity;
-        JsonProject jsonProject = null;
+        Project projectEntity = this.projectService.findByNameOrId( name );
 
-        try {
-            long id = Long.parseLong(name);
-            projectEntity = this.projectService.find(id);
-        } catch (NumberFormatException nfe) {
-            LOG.warn("Value " + name + " could not be parsed into a number. Trying to find the project by name.");
-            projectEntity = this.projectService.findByName(name);
+        if ( null != projectEntity ) {
+            JsonProject jsonProject = this.projectService.convertEntityToJson( projectEntity );
+            return Response.status( Response.Status.OK ).entity( jsonProject ).build();
         }
 
-        if (null != projectEntity) {
-            jsonProject = this.projectService.convertEntityToJson(projectEntity);
-        }
-
-        return jsonProject;
+        return Response.status( Response.Status.NO_CONTENT ).build();
     }
 
     /**
@@ -116,51 +102,58 @@ public class ProjectResource
      * @return List<Projects> projects
      */
     @GET
-    @Path("directsearch")
-    @ApiOperation(value = "Finds a list of projects in the repository by search query. <BR>"
-            + "The search query has to be submitted in the Elastic Search JSON search syntax.")
-    @ApiResponses(value = {
-            @ApiResponse(code = HttpStatus.SC_NO_CONTENT, message = "No content"),
-            @ApiResponse(code = HttpStatus.SC_BAD_REQUEST, message = "Request incorrect"),
-            @ApiResponse(code = HttpStatus.SC_NOT_FOUND, message = "Not found"),
-            @ApiResponse(code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error")})
-    @Produces(MediaType.APPLICATION_JSON)
-    public List<ProjectDTO> searchProjects(@QueryParam("q")
-                                           @ApiParam(value = "Search query in the Elastic Search JSON syntax") final String query)
-    {
-        LOG.info("search '{}'", query);
-        final List<ProjectDTO> projects = this.projectService.find(query);
-        return projects;
+    @Path( "directsearch" )
+    @ApiOperation( value = "Finds a list of projects in the repository by search query. <BR>"
+            + "The search query has to be submitted in the Elastic Search JSON search syntax.",
+            response = ProjectDTO.class,
+            responseContainer = "List" )
+    @ApiResponses( value = {
+            @ApiResponse( code = HttpStatus.SC_OK, message = "OK" ),
+            @ApiResponse( code = HttpStatus.SC_NO_CONTENT, message = "No content" ),
+            @ApiResponse( code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error" )} )
+    @Produces( MediaType.APPLICATION_JSON )
+    public Response searchProjects( @QueryParam( "q" )
+                                    @ApiParam( value = "Search query in the Elastic Search JSON syntax" ) final String query ) {
+        LOG.info( "search '{}'", query );
+        final List<ProjectDTO> projects = this.projectService.find( query );
+        if ( projects.isEmpty() ) {
+            return Response.status( Response.Status.NO_CONTENT ).build();
+        }
+        return Response.status( Response.Status.OK ).entity( projects ).build();
     }
 
 
     /**
      * Find a list of projects by free text search strings.
      *
-     * @param query the query
-     * @param domain the domain
+     * @param query     the query
+     * @param domain    the domain
      * @param subdomain the sub domain
      * @return List<ProjectsDTO> projects
      */
     @GET
-    @Path("search")
-    @ApiOperation(value = "Finds a list of projects by free text search strings. <BR> "
-            + "All query parameters are optional and can be combined as needed but at least one parameter has to be submitted.")
-    @ApiResponses(value = {
-            @ApiResponse(code = HttpStatus.SC_NO_CONTENT, message = "No content"),
-            @ApiResponse(code = HttpStatus.SC_BAD_REQUEST, message = "Request incorrect"),
-            @ApiResponse(code = HttpStatus.SC_NOT_FOUND, message = "Not found"),
-            @ApiResponse(code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error")})
-    @Produces(MediaType.APPLICATION_JSON)
-    public List<ProjectDTO> searchProjects(
-            @QueryParam("query") @ApiParam(value = "Free text string that should be used to search inside of a project")
-                final String query,
-            @QueryParam("domain") @ApiParam(value = "Domain name string a project should be assigned to") final String domain,
-            @QueryParam("subdomain") @ApiParam(value = "Subdomain name string a project should be assigned to")
-                final String subdomain)
-    {
-        final List<ProjectDTO> projects = this.projectService.find(query, domain, subdomain);
-        return projects;
+    @Path( "search" )
+    @ApiOperation( value = "Finds a list of projects by free text search strings. <BR> "
+            + "All query parameters are optional and can be combined as needed but at least one parameter has to be submitted.",
+            response = ProjectDTO.class,
+            responseContainer = "List" )
+    @ApiResponses( value = {
+            @ApiResponse( code = HttpStatus.SC_OK, message = "OK" ),
+            @ApiResponse( code = HttpStatus.SC_NO_CONTENT, message = "No content" ),
+            @ApiResponse( code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error" )} )
+    @Produces( MediaType.APPLICATION_JSON )
+    public Response searchProjects(
+            @QueryParam( "query" ) @ApiParam( value = "Free text string that should be used to search inside of a project" )
+            final String query,
+            @QueryParam( "domain" ) @ApiParam( value = "Domain name string a project should be assigned to" ) final String domain,
+            @QueryParam( "subdomain" ) @ApiParam( value = "Subdomain name string a project should be assigned to" )
+            final String subdomain ) {
+        final List<ProjectDTO> projects = this.projectService.find( query, domain, subdomain );
+
+        if ( projects.isEmpty() ) {
+            return Response.status( Response.Status.NO_CONTENT ).build();
+        }
+        return Response.status( Response.Status.OK ).entity( projects ).build();
     }
 
     /**
@@ -170,18 +163,16 @@ public class ProjectResource
      * @return {#link javax.ws.rs.core.Response Response}
      */
     @POST
-    @ApiOperation(value = "Creates and stores a new project in the repository")
-    @ApiResponses(value = {
-            @ApiResponse(code = HttpStatus.SC_NO_CONTENT, message = "No content (project has been created successfully)"),
-            @ApiResponse(code = HttpStatus.SC_BAD_REQUEST, message = "Request incorrect"),
-            @ApiResponse(code = HttpStatus.SC_NOT_FOUND, message = "Not found"),
-            @ApiResponse(code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error")})
-    public Response create(@ApiParam(value = "The project to be stored inside of the Asset Repo") JsonProject project)
-    {
-        final Project projectEntity = this.projectService.convertJsonToEntity(project);
-        final Project created = this.projectService.create(projectEntity);
+    @ApiOperation( value = "Creates and stores a new project in the repository" )
+    @ApiResponses( value = {
+            @ApiResponse( code = HttpStatus.SC_CREATED, message = "Created (project has been created successfully)" ),
+            @ApiResponse( code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error" )} )
+    public Response create( @ApiParam( value = "The project to be stored inside of the Asset Repo" ) JsonProject project ) {
 
-        return redirect("project/" + created.getId());
+        final Project projectEntity = this.projectService.convertJsonToEntity( project );
+        final Project created = this.projectService.create( projectEntity );
+        final JsonProject jsonProject = this.projectService.convertEntityToJson( created );
+        return Response.status( Response.Status.CREATED ).entity( jsonProject ).build();
     }
 
     /**
@@ -192,20 +183,24 @@ public class ProjectResource
      * @return {#link javax.ws.rs.core.Response HTTP Response code}
      */
     @PUT
-    @Path("{id}")
-    @ApiOperation(value = "Updates a project in the Asset Repo")
-    @ApiResponses(value = {
-            @ApiResponse(code = HttpStatus.SC_NO_CONTENT, message = "No content"),
-            @ApiResponse(code = HttpStatus.SC_BAD_REQUEST, message = "Request incorrect"),
-            @ApiResponse(code = HttpStatus.SC_NOT_FOUND, message = "Not found"),
-            @ApiResponse(code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error")})
-    public Response update(@PathParam("id") @ApiParam(value = "Project ID") long id,
-                           @ApiParam(value = "The project to be updated inside of the Asset Repo") JsonProject project)
-    {
-        project.setId(id);
-        final Project projectEntity = this.projectService.convertJsonToEntity(project);
-        final Project updated = this.projectService.update(projectEntity);
-        return redirect("project/", updated);
+    @Path( "{id}" )
+    @ApiOperation( value = "Updates a project in the Asset Repo" )
+    @ApiResponses( value = {
+            @ApiResponse( code = HttpStatus.SC_OK, message = "OK" ),
+            @ApiResponse( code = HttpStatus.SC_NOT_FOUND, message = "Not found" ),
+            @ApiResponse( code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error" )} )
+    public Response update( @PathParam( "id" ) @ApiParam( value = "Project ID" ) long id,
+                            @ApiParam( value = "The project to be updated inside of the Asset Repo" ) JsonProject project ) {
+
+        if ( null != projectService.find( id ) ) {
+            project.setId( id );
+            final Project projectEntity = this.projectService.convertJsonToEntity( project );
+            final Project updated = this.projectService.update( projectEntity );
+            final JsonProject jsonProject = this.projectService.convertEntityToJson( updated );
+            return Response.status( Response.Status.OK ).entity( jsonProject ).build();
+        }
+
+        return Response.status( Response.Status.NOT_FOUND ).build();
     }
 
     /**
@@ -214,16 +209,18 @@ public class ProjectResource
      * @param id ID of the project to be deleted
      */
     @DELETE
-    @Path("{id}")
-    @ApiOperation(value = "Deletes a project from the Asset Repo by ID")
-    @ApiResponses(value = {
-            @ApiResponse(code = HttpStatus.SC_NO_CONTENT, message = "No content"),
-            @ApiResponse(code = HttpStatus.SC_BAD_REQUEST, message = "Request incorrect"),
-            @ApiResponse(code = HttpStatus.SC_NOT_FOUND, message = "Not found"),
-            @ApiResponse(code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error")})
-    public void delete(@PathParam("id") @ApiParam(value = "ID of the project to be deleted") long id)
-    {
-        this.projectService.delete(id);
+    @Path( "{id}" )
+    @ApiOperation( value = "Deletes a project from the Asset Repo by ID" )
+    @ApiResponses( value = {
+            @ApiResponse( code = HttpStatus.SC_OK, message = "OK" ),
+            @ApiResponse( code = HttpStatus.SC_NOT_FOUND, message = "Not found" ),
+            @ApiResponse( code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error" )} )
+    public Response delete( @PathParam( "id" ) @ApiParam( value = "ID of the project to be deleted" ) long id ) {
+        if ( null != projectService.find( id ) ) {
+            this.projectService.delete( id );
+            return Response.status( Response.Status.OK ).build();
+        }
+        return Response.status( Response.Status.NOT_FOUND ).build();
     }
 
     /**
@@ -232,15 +229,19 @@ public class ProjectResource
      * @param name Name string of the project to be deleted
      */
     @DELETE
-    @Path("{name}")
-    @ApiOperation(value = "Deletes a project from the Asset Repo by name")
-    @ApiResponses(value = {
-            @ApiResponse(code = HttpStatus.SC_NO_CONTENT, message = "No content"),
-            @ApiResponse(code = HttpStatus.SC_BAD_REQUEST, message = "Request incorrect"),
-            @ApiResponse(code = HttpStatus.SC_NOT_FOUND, message = "Not found"),
-            @ApiResponse(code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error")})
-    public void delete(@PathParam("name") @ApiParam(value = "name of the project to be deleted") String name)
-    {
-        this.projectService.delete(name);
+    @Path( "{name}" )
+    @ApiOperation( value = "Deletes a project from the Asset Repo by name" )
+    @ApiResponses( value = {
+            @ApiResponse( code = HttpStatus.SC_OK, message = "OK" ),
+            @ApiResponse( code = HttpStatus.SC_NOT_FOUND, message = "Not found" ),
+            @ApiResponse( code = HttpStatus.SC_INTERNAL_SERVER_ERROR, message = "Internal Server error" )} )
+    public Response delete( @PathParam( "name" ) @ApiParam( value = "name of the project to be deleted" ) String name ) {
+
+        Project project = projectService.findByNameOrId( name );
+        if ( null != project ) {
+            this.projectService.delete( project );
+            return Response.status( Response.Status.OK ).build();
+        }
+        return Response.status( Response.Status.NOT_FOUND ).build();
     }
 }
